Run render & mount actions when components are mounted
When is a component mounted (aside from the root component, which is
explicitly mounted)? Perhaps when its last child is mounted?

This takes the approach of setting a snabbdom insert hook on each
component's last child. These hooks run the render actions then the
mount actions.

This *seems* to be working, but it feels wonky and there are likely to
be bugs.

In addition, this also makes registration of render and mount actions
a little nicer by adding some utility functions.