Switch to an Elm-like architecture
Which I guess is also an Redux-like architecture.

I think this is a good way to manage state vs components maintaining
their own state and having to pass `set` functions around. This also
makes the internals a bit simpler.

The one major issue is that rerendering isn't currently triggering
on-render actions. Still trying to figure the best way to determine when
a component is rerendered.