Run component render actions on rerender as well as first render
Previously, render actions were run on mount/first render because they
were triggered explicitly from `mount()`, but they weren't being called
on rerender because there wasn't a way to tell when a component was
rerendered.

This adds a mechanism for determining when a component is rendered and
running its render actions, handling the first render and rerenders in
the same way.

A component is considered to have been rendered if one of its descendent
nodes is added or removed (excluding children of subcomponents). In
addition, if the content of a text node is updated, that's considered
a rerender too.

The way this is accomplished feels a little wonky, but it seems to work
okay.

This also changes the order of mount and render actions due to how
render detection is implemented. Previously, render actions would be run
before mount actions, but now mount actions run first.